requirements.txt
scikit-learn
streamlit
pdfplumber
python-docx
openai
sentence-transformers
torch
python-dotenv
